requirements.txt
decorator==4.0.4
Django==1.9
django-admin-bootstrapped==2.5.6
djangorestframework==3.3.1
ipdb==0.8.1
ipython==4.0.1
ipython-genutils==0.1.0
mysqlclient==1.3.7
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
prettyconf==1.2.0
ptyprocess==0.5
requests==2.7.0
simplegeneric==0.8.1
traitlets==4.0.0
wheel==0.24.0
